--- js/get-random-element.js
@@ -1,6 +1,0 @@
-//import _ from ('lodash');
-const getRandomElement = (elementsArray) => {
-  return _.sample([elementsArray]);
-};
-
-export {getRandomElement};
