+++ js/main.js
@@ -1,5 +1,6 @@
-import {getCardsArray} from './get-cards-array.js';
+import {getCardsArray} from './data.js';
 
 const FINAL_ARRAY_ELEMENTS = 10;
 
 getCardsArray(FINAL_ARRAY_ELEMENTS);
+console.log(getCardsArray(FINAL_ARRAY_ELEMENTS));
